+++ @ecomplus/storefront-components/src/js/helpers/add-idle-callback.js
@@ -1,0 +1,7 @@
+export default fn => {
+  if (typeof window === 'object' && typeof window.requestIdleCallback === 'function') {
+    window.requestIdleCallback(fn)
+  } else {
+    setTimeout(fn, 500)
+  }
+}