+++ @ecomplus/storefront-template/template/js/lib/clearfix.js
@@ -1,0 +1,7 @@
+import ecomCart from '@ecomplus/shopping-cart'
+
+ecomCart.data.items.forEach(({ _id, flags }) => {
+  if (Array.isArray(flags) && flags.includes('__tmp')) {
+    ecomCart.removeItem(_id)
+  }
+})