+++ @ecomplus/storefront-template/template/js/netlify-cms/base-config/collections/pages/search.js
@@ -22,7 +22,7 @@ export default ({ baseDir, sections }) => ({
             }
           ]
         }
-      ].concat(sections)
+      ].concat(sections.filter(({ name }) => name !== 'collection-shelf'))
     }
   ]
 })